requirements.txt
flask
pytest
azure-data-tables
